final update

+++ app/Http/Controllers/ArsipController.php
@@ -17,125 +17,100 @@ class ArsipController extends Controller
 
     public function index()
     {
-        // Di Controller Anda
-        $arsipsMasuk = Arsip::where('jenis_arsip', 'masuk')->get();
-        $arsipsKeluar = Arsip::where('jenis_arsip', 'keluar')->get();
-        $arsips = Arsip::with('users')->get();
+        if (auth()->user()->hasRole('user')) { // Ambil semua arsip beserta relasi users dan data pivot
+            $userId = auth()->id(); // Mendapatkan ID user yang sedang login
+            // Ambil arsip yang memiliki relasi dengan user yang sedang login
+            $arsips = Arsip::whereHas('users', function ($query) use ($userId) {
+                $query->where('users.id', $userId);
+            })->with('users')->get();
+        } else {
+            $arsips = Arsip::with('users')->get();
+        }
+        // Jika Anda ingin memisahkan berdasarkan jenis arsip (masuk/keluar), Anda bisa melakukan:
+        $arsipsMasuk = $arsips->where('jenis_arsip', 'masuk');
+        $arsipsKeluar = $arsips->where('jenis_arsip', 'keluar');
+
         return view('admin.arsip.index', compact('arsips', 'arsipsMasuk', 'arsipsKeluar'));
     }
 
-    public function create()
+    public function create(Request $request)
     {
+        $jenis_arsip = $request->query('jenis_arsip');
+        if (auth()->user()->hasRole('direktur')) {
+            return redirect()->route(getCurrentRoutePrefix() . '.arsip.index')->with('error', 'Anda tidak memiliki izin untuk melakukan aksi ini.');
+        }
         $klasifikasiArsips = KlasifikasiArsip::all();
         $lemaris = Lemari::all();
         $raks = Rak::all();
         $folders = Folder::all();
-        $users = User::where('role_id', '4')->get();
-        return view('admin.arsip.create', compact('klasifikasiArsips', 'lemaris', 'raks', 'folders', 'users'));
+        $users = User::whereIn('role_id', [3, 4])->get();
+        return view('admin.arsip.create', compact('jenis_arsip', 'klasifikasiArsips', 'lemaris', 'raks', 'folders', 'users'));
     }
 
     public function store(Request $request)
     {
-        // Validasi data yang masuk
-
-        $validatedData = $request->validate([
+        $jenisArsip = $request->input('jenis_arsip');
+        // Atur validasi berdasarkan jenis arsip
+        $rules = [
             'nomor_surat' => 'required|string|max:50',
+            'hal' => 'required|string|max:255',
             'jumlah' => 'required|integer',
-            'dari' => 'required|string|max:50',
-            'kepada' => 'required|array',
-            'kepada.*' => 'required|integer|exists:users,id',
             'sifat' => 'required|in:rahasia,biasa,segera,sangat segera',
             'jenis_arsip' => 'required|in:masuk,keluar',
             'keamanan_arsip' => 'required|in:asli,fotocopy',
             'lampiran' => 'sometimes|file|mimes:jpeg,png,jpg,pdf,docx|max:2048',
-            'keterangan' => 'required|string|max:255',
+            'keterangan' => 'nullable|string|max:255',
             'tanggal_arsip' => 'required|date',
             'status_arsip' => 'required|in:diproses,selesai,palsu,meragukan,disposisi',
-
             'klasifikasi_id' => 'required|exists:klasifikasi_arsip,id_klasifikasi_arsip',
             'lemari_id' => 'required|exists:lemari,id_lemari',
             'rak_id' => 'required|exists:rak,id_rak',
             'folder_id' => 'required|exists:folder,id_folder',
-        ], [
-            // Custom error messages
-            'nomor_surat.required' => 'Nomor surat wajib diisi.',
-            'nomor_surat.string' => 'Nomor surat harus berupa teks.',
-            'nomor_surat.max' => 'Nomor surat tidak boleh lebih dari 50 karakter.',
+        ];
 
-            'jumlah.required' => 'Jumlah wajib diisi.',
-            'jumlah.integer' => 'Jumlah harus berupa angka.',
+        if ($jenisArsip === 'keluar') {
+            $rules['kepada'] = 'required|string|max:50';
+            $rules['dari'] = 'required|array';
+            $rules['dari.*'] = 'required|integer|exists:users,id';
+        } else {
+            $rules['dari'] = 'required|string|max:50';
+            $rules['kepada'] = 'required|array';
+            $rules['kepada.*'] = 'required|integer|exists:users,id';
+        }
 
-            'dari.required' => 'Field Dari wajib diisi.',
-            'dari.string' => 'Field Dari harus berupa teks.',
-            'dari.max' => 'Field Dari tidak boleh lebih dari 50 karakter.',
-
-            'kepada.required' => 'Field Kepada wajib diisi.',
-            'kepada.array' => 'Field Kepada harus berupa array.',
-
-
-            'sifat.required' => 'Sifat wajib diisi.',
-            'sifat.in' => 'Sifat harus salah satu dari: rahasia, biasa, segera, sangat segera.',
-
-            'jenis_arsip.required' => 'Jenis Arsip wajib diisi.',
-            'jenis_arsip.in' => 'Jenis Arsip harus salah satu dari: masuk, keluar.',
-
-            'keamanan_arsip.required' => 'Keamanan Arsip wajib diisi.',
-            'keamanan_arsip.in' => 'Keamanan Arsip harus salah satu dari: asli, fotocopy.',
-
-            'lampiran.required' => 'Lampiran wajib diisi.',
-
-            'keterangan.required' => 'Keterangan wajib diisi.',
-            'keterangan.string' => 'Keterangan harus berupa teks.',
-            'keterangan.max' => 'Keterangan tidak boleh lebih dari 255 karakter.',
-
-            'tanggal_arsip.required' => 'Tanggal Arsip wajib diisi.',
-            'tanggal_arsip.date' => 'Format tanggal tidak valid.',
-
-            'status_arsip.required' => 'Status Arsip wajib diisi.',
-            'status_arsip.in' => 'Status Arsip harus salah satu dari: diproses, selesai, palsu, meragukan, disposisi.',
-
-            'klasifikasi_id.required' => 'Klasifikasi ID wajib diisi.',
-            'klasifikasi_id.exists' => 'Klasifikasi ID tidak valid.',
-
-            'lemari_id.required' => 'Lemari ID wajib diisi.',
-            'lemari_id.exists' => 'Lemari ID tidak valid.',
-
-            'rak_id.required' => 'Rak ID wajib diisi.',
-            'rak_id.exists' => 'Rak ID tidak valid.',
-
-            'folder_id.required' => 'Folder ID wajib diisi.',
-            'folder_id.exists' => 'Folder ID tidak valid.',
-        ]);
+        // Lakukan validasi
+        $validatedData = $request->validate($rules);
 
         // Jika ada file lampiran yang diupload
+        $filename = null;
         if ($request->hasFile('lampiran')) {
             $file = $request->file('lampiran');
             $filename = time() . '.' . $file->getClientOriginalExtension();
             $file->storeAs('public/lampiran', $filename);
-        } else {
-            $filename = null;
         }
 
         // Menyimpan data ke database
-        $arsip = new Arsip;
-        $arsip->nomor_surat = $validatedData['nomor_surat'];
-        $arsip->jumlah = $validatedData['jumlah'];
-        $arsip->dari = $validatedData['dari'];
-        $arsip->sifat = $validatedData['sifat'];
+        $arsip = new Arsip($validatedData);
+        $arsip->lampiran = $filename;
         $arsip->jenis_arsip = $validatedData['jenis_arsip'];
         $arsip->keamanan_arsip = $validatedData['keamanan_arsip'];
-        $arsip->lampiran = $filename;
-        $arsip->keterangan = $validatedData['keterangan'];
-        $arsip->tanggal_arsip = $validatedData['tanggal_arsip'];
         $arsip->status_arsip = $validatedData['status_arsip'];
-
         $arsip->klasifikasi_id = $validatedData['klasifikasi_id'];
         $arsip->lemari_id = $validatedData['lemari_id'];
         $arsip->rak_id = $validatedData['rak_id'];
         $arsip->folder_id = $validatedData['folder_id'];
-        $arsip->save();
+        $arsip->tanggal_arsip = $validatedData['tanggal_arsip'];
+        $arsip->keterangan = $validatedData['keterangan'];
 
-        $arsip->users()->attach($request->kepada);
+        if ($jenisArsip === 'keluar') {
+            $arsip->dari = $request->input('kepada');
+            $arsip->save();
+            $arsip->users()->attach($request->input('dari'));
+        } else {
+            $arsip->dari = $validatedData['dari'];
+            $arsip->save();
+            $arsip->users()->attach($request->input('kepada'));
+        }
 
         // Redirect ke halaman daftar arsip dengan pesan sukses
         return redirect()->route(getCurrentRoutePrefix() . '.arsip.index')->with('success', 'Arsip berhasil ditambahkan.');
@@ -158,7 +133,7 @@ class ArsipController extends Controller
         $raks = Rak::all();
         $folders = Folder::all();
         $arsip = Arsip::findOrFail($id);
-        $users = User::where('role_id', '4')->get();
+        $users = User::whereIn('role_id', [3, 4])->get();
         $selectedUsers = $arsip->users->pluck('id')->toArray();
         return view('admin.arsip.edit', compact('arsip', 'klasifikasiArsips', 'lemaris', 'raks', 'folders', 'users', 'selectedUsers'));
     }
@@ -166,9 +141,10 @@ class ArsipController extends Controller
     public function update(Request $request, $id) // $id bertipe data sesuai dengan tipe id di database
     {
         $arsip = Arsip::findOrFail($id);
-        
+
         $data = $request->validate([
             'nomor_surat' => 'required|string|max:50',
+            'hal' => 'required|string|max:255',
             'jumlah' => 'required|integer',
             'dari' => 'required|string|max:50',
             'kepada' => 'required|array',
@@ -177,7 +153,7 @@ class ArsipController extends Controller
             'jenis_arsip' => 'required|in:masuk,keluar',
             'keamanan_arsip' => 'required|in:asli,fotocopy',
             'lampiran' => 'sometimes|file|mimes:jpeg,png,jpg,pdf,docx|max:2048',
-            'keterangan' => 'required|string|max:255',
+            'keterangan' => 'nullable|string|max:255',
             'tanggal_arsip' => 'required|date',
             'status_arsip' => 'required|in:diproses,selesai,palsu,meragukan,disposisi',
             'users_id' => 'required|exists:users,id',
@@ -190,6 +166,10 @@ class ArsipController extends Controller
             'nomor_surat.required' => 'Nomor surat wajib diisi.',
             'nomor_surat.string' => 'Nomor surat harus berupa teks.',
             'nomor_surat.max' => 'Nomor surat tidak boleh lebih dari 50 karakter.',
+
+            'hal.required' => 'Kolom hal harus diisi.',
+            'hal.string' => 'Kolom hal harus berupa teks.',
+            'hal.max' => 'Kolom hal tidak boleh lebih dari 255 karakter.',
 
             'jumlah.required' => 'Jumlah wajib diisi.',
             'jumlah.integer' => 'Jumlah harus berupa angka.',
@@ -212,7 +192,6 @@ class ArsipController extends Controller
 
             'lampiran.required' => 'Lampiran wajib diisi.',
 
-            'keterangan.required' => 'Keterangan wajib diisi.',
             'keterangan.string' => 'Keterangan harus berupa teks.',
             'keterangan.max' => 'Keterangan tidak boleh lebih dari 255 karakter.',
 
@@ -241,7 +220,7 @@ class ArsipController extends Controller
 
         // Jika ada file lampiran yang diupload
         if ($request->hasFile('lampiran')) {
-            // Hapus gambar lama jika ada
+            // Hapus lampiran lama jika ada
             if ($arsip->lampiran) {
                 Storage::delete('public/lampiran/' . $arsip->lampiran);
             }
@@ -262,7 +241,11 @@ class ArsipController extends Controller
 
     public function destroy($id)
     {
+
         try {
+            if (auth()->user()->hasRole('direktur')) {
+                return redirect()->route(getCurrentRoutePrefix() . '.disposisi.index')->with('error', 'Anda tidak memiliki izin untuk melakukan aksi ini.');
+            }
             $arsip = Arsip::findOrFail($id);
             $arsip->users()->detach();
             $arsip->delete();
@@ -278,4 +261,47 @@ class ArsipController extends Controller
             return redirect()->route(getCurrentRoutePrefix() . '.arsip.index')->with('error', 'Terjadi kesalahan saat menghapus arsip.');
         }
     }
+
+
+    public function acceptArsip($id)
+    {
+        $arsip = Arsip::findOrFail($id);
+        $userId = auth()->id();
+
+        // Cek status sebelumnya
+        $existingStatus = $arsip->users()->where('user_id', $userId)->first();
+        if ($existingStatus && $existingStatus->pivot->status !== 'perlu konfirmasi') {
+            return redirect()->route(getCurrentRoutePrefix() . '.arsip.index')->with('error', 'Anda sudah mengonfirmasi status arsip ini sebelumnya.');
+        }
+
+        // Update status di tabel pivot arsip
+        $arsip->users()->updateExistingPivot($userId, ['status' => 'diterima']);
+        return redirect()->route(getCurrentRoutePrefix() . '.arsip.index')->with('success', 'Status arsip berhasil diubah menjadi diterima.');
+    }
+
+    public function declineArsip(Request $request, $id)
+    {
+        $arsip = Arsip::findOrFail($id);
+        $userId = auth()->user()->id;
+        $keteranganDisposisi = $request->input('keterangan_disposisi');
+
+        // Cek status sebelumnya
+        $existingStatus = $arsip->users()->where('user_id', $userId)->first();
+        if ($existingStatus && $existingStatus->pivot->status !== 'perlu konfirmasi') {
+            return redirect()->route(getCurrentRoutePrefix() . '.arsip.index')
+                ->with('error', 'Anda sudah mengonfirmasi status arsip ini sebelumnya, hanya bisa 1x perubahan!');
+        }
+
+        // Update status dan keterangan disposisi di tabel pivot
+        $arsip->users()->updateExistingPivot($userId, [
+            'status' => 'disposisi',
+            'disposisi_keterangan' => $keteranganDisposisi // Menyimpan keterangan
+        ]);
+        $arsip->update(['status_arsip' => 'disposisi']);
+
+        return redirect()->route(getCurrentRoutePrefix() . '.arsip.index')
+            ->with('success', 'Status arsip berhasil diubah menjadi disposisi dengan keterangan.');
+    }
+
+    
 }
